requirements.txt
wandb>=0.18.7
transformers>=4.46.3
datasets>=3.1.0
PyYAML>=6.0.2
torch>=2.5.1
evaluate>=0.4.3
gradio>=5.7.1
newspaper3k>=0.2.8
lxml_html_clean>=0.4.1
accelerate>=1.1.1
absl-py>=2.1.0
rouge_score>=0.1.2
pytest>=8.3.4
pytest-cov>=6.0.0
nltk~=3.9.1